+++ lua/rustaceanvim/config/init.lua
@@ -85,6 +85,7 @@ vim.g.rustaceanvim = vim.g.rustaceanvim
 ---@field cmd? string[] | fun():string[] Command and arguments for starting rust-analyzer
 ---@field settings? fun(project_root:string|nil):table | table Setting passed to rust-analyzer. Defaults to a function that looks for a `rust-analyzer.json` file or returns an empty table. See https://rust-analyzer.github.io/manual.html#configuration.
 ---@field standalone? boolean Standalone file support (enabled by default). Disabling it may improve rust-analyzer's startup time.
+---@field logfile? string The path to the rust-analyzer log file.
 
 ---@class RustaceanDapOpts
 ---@field adapter? DapExecutableConfig | DapServerConfig | disable | fun():(DapExecutableConfig | DapServerConfig | disable) Defaults to a `DapServerConfig` if `codelldb` is detected, and to a `DapExecutableConfig` if `lldb` is detected. Set to `false` to disable.